fix: lint errors in plugin-loader.ts
- Remove unused serverLogger import
- Tag TODO with issue reference
- Remove async from functions that don't use await
